requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
mediapipe>=0.8.0
pyyaml>=6.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.21.0
tqdm>=4.66.0
python-dotenv>=1.0.0
